pyproject.toml
[tool.poetry]
name = "mktdata-collector"
version = "0.1.0"
description = ""
authors = ["Guilherme Rodrigues <55053188+guilhermegor@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "1.0.1"
pre-commit = "^4.0.1"
workalendar = "15.0.1"
businesstimedelta = "1.0.1"
pandas = "2.1.2"
pathlib = "1.0.1"
numpy = ">=1.26.0,<2.0.0"
requests = "2.31.0"
urllib3 = "1.26.3"
lxml = "4.9.3"
psycopg2 = "2.8.6"
stpstone = "0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
